+++ src/library/occupancy_grid_map/point_cloud_obstacle.h
@@ -54,8 +54,8 @@ class PointCloudObstacleTransform {
 
   void GenerateLocalObstacleByLocalView(ParkObstacleList& obs_list,
                                         const LocalView* local_view,
-                                        const double slot_length,
-                                        const double slot_width,
+                                        const float slot_length,
+                                        const float slot_width,
                                         const Pose2D& slot_base_pose,
                                         const Pose2D& ego_start,
                                         const bool enable_limiter_obs);
@@ -69,8 +69,8 @@ class PointCloudObstacleTransform {
       ParkObstacleList& obs_list, cdl::AABB& box);
 
  private:
-  void SampleInLineSegment(const Eigen::Vector2d& start,
-                           const Eigen::Vector2d& end,
+  void SampleInLineSegment(const Eigen::Vector2f& start,
+                           const Eigen::Vector2f& end,
                            std::vector<Position2D>* points);
 };
